firmware: use int16_t for recv_fps's return value
This fixes a bug with framerates above 127.

+++ firmware/firmware.ino
@@ -92,7 +92,7 @@ setup()
 void
 loop()
 {
-    int8_t ret;
+    int16_t ret;
     uint32_t i, t;
 
     /* Wait for once-a-second packet. */
